+++ db/filmorate.trace.db
@@ -7093,3 +7093,15 @@ java.sql.SQLClientInfoException: Client info name 'ApplicationName' not supporte
 	at java.base/java.util.concurrent.ThreadPoolExecutor.runWorker(ThreadPoolExecutor.java:1136)
 	at java.base/java.util.concurrent.ThreadPoolExecutor$Worker.run(ThreadPoolExecutor.java:635)
 	at java.base/java.lang.Thread.run(Thread.java:833)
+2022-11-15 22:19:36 jdbc[3]: exception
+org.h2.jdbc.JdbcSQLSyntaxErrorException: Таблица "RATING" не найдена
+Table "RATING" not found; SQL statement:
+MERGE INTO RATING KEY (ID) VALUES ( 1, 'G' ), ( 2, 'PG' ), ( 3, 'PG13' ), ( 4, 'R' ), ( 5, 'NC17' ) [42102-214]
+2022-11-15 22:21:45 jdbc[3]: exception
+org.h2.jdbc.JdbcSQLSyntaxErrorException: Таблица "RATING" не найдена
+Table "RATING" not found; SQL statement:
+MERGE INTO RATING KEY (ID) VALUES ( 1, 'G' ), ( 2, 'PG' ), ( 3, 'PG13' ), ( 4, 'R' ), ( 5, 'NC17' ) [42102-214]
+2022-11-15 22:47:32 jdbc[3]: exception
+org.h2.jdbc.JdbcSQLSyntaxErrorException: Таблица "RATING" не найдена
+Table "RATING" not found; SQL statement:
+MERGE INTO RATING KEY (ID) VALUES ( 1, 'G' ), ( 2, 'PG' ), ( 3, 'PG13' ), ( 4, 'R' ), ( 5, 'NC17' ) [42102-214]
