+++ src/main/resources/schema.sql
@@ -1,9 +1,22 @@
+CREATE ALIAS IF NOT EXISTS getDate AS
+'java.util.Date getDate() {
+    return new java.util.Date();
+}';
+
 CREATE TABLE IF NOT EXISTS users (
     id integer GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
     email varchar NOT NULL UNIQUE,
     login varchar NOT NULL UNIQUE,
     name varchar,
-    birthday date
+    birthday date,
+    CONSTRAINT constraint_login
+        CHECK (login NOT LIKE '% %' AND login NOT LIKE ' '),
+    CONSTRAINT constraint_email
+        check (email like '%@%'),
+    CONSTRAINT constraint_birthday
+            CHECK (CAST(birthday AS date) <= CAST(getDate() as date))
+
+
 );
 
 CREATE TABLE IF NOT EXISTS follow (
@@ -13,12 +26,12 @@ CREATE TABLE IF NOT EXISTS follow (
 );
 
 CREATE TABLE IF NOT EXISTS genre (
-    id integer GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
+    id integer PRIMARY KEY,
     name varchar NOT NULL
 );
 
 CREATE TABLE IF NOT EXISTS MPA (
-    id integer GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
+    id integer NOT NULL PRIMARY KEY,
     name varchar NOT NULL
 );
 
@@ -26,10 +39,14 @@ CREATE TABLE IF NOT EXISTS movie (
     id integer GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
     name varchar NOT NULL,
     description varchar(200),
-    release_date date,
-    duration integer,
-    rate integer,
-    mpa_id integer REFERENCES MPA (id)
+    release_date date NOT NULL,
+    duration integer NOT NULL,
+    rate integer NOT NULL,
+    mpa_id integer REFERENCES MPA (id),
+    CONSTRAINT constraint_release_date
+    check (release_date > '1895-12-27'),
+    CONSTRAINT "movie_mpa_id" FOREIGN KEY (mpa_id) REFERENCES mpa (id)
+
 );
 
 CREATE TABLE IF NOT EXISTS film_genre (
